Refactor to allow metrics only collection
